pyproject.toml
[project]
name = "generative-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.17",
    "langchain-core>=1.3.2",
    "langchain-gemini>=0.1.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
    "streamlit>=1.57.0",
]
